Move Lerps and Easing functions
Also fix Bezier factory so it works with Tween
